add results to import
