+++ packages/core/strapi/lib/commands/transfer/export.js
@@ -7,30 +7,12 @@ const {
   // TODO: we need to solve this issue with typescript modules
   // eslint-disable-next-line import/no-unresolved, node/no-missing-require
 } = require('@strapi/data-transfer');
-const _ = require('lodash/fp');
-const Table = require('cli-table3');
+const { isObject, isString, isFinite } = require('lodash/fp');
 const fs = require('fs-extra');
 
 const chalk = require('chalk');
 const strapi = require('../../index');
-const { readableBytes } = require('../utils');
-
-const pad = (n) => {
-  return (n < 10 ? '0' : '') + String(n);
-};
-
-const yyyymmddHHMMSS = () => {
-  const date = new Date();
-
-  return (
-    date.getFullYear() +
-    pad(date.getMonth() + 1) +
-    pad(date.getDate()) +
-    pad(date.getHours()) +
-    pad(date.getMinutes()) +
-    pad(date.getSeconds())
-  );
-};
+const { buildTransferTable, yyyymmddHHMMSS } = require('./util');
 
 const getDefaultExportName = () => {
   return `export_${yyyymmddHHMMSS()}`;
@@ -42,7 +24,7 @@ const BYTES_IN_MB = 1024 * 1024;
 
 module.exports = async (opts) => {
   // validate inputs from Commander
-  if (!_.isObject(opts)) {
+  if (!isObject(opts)) {
     logger.error('Could not parse arguments');
     process.exit(1);
   }
@@ -58,7 +40,7 @@ module.exports = async (opts) => {
   };
   const source = createLocalStrapiSourceProvider(sourceOptions);
 
-  const file = _.isString(filename) && filename.length > 0 ? filename : getDefaultExportName();
+  const file = isString(filename) && filename.length > 0 ? filename : getDefaultExportName();
 
   /**
    * To a Strapi backup file
@@ -67,8 +49,8 @@ module.exports = async (opts) => {
   const destinationOptions = {
     file: {
       path: file,
-      maxSize: _.isFinite(opts.maxSize) ? Math.floor(opts.maxSize) * BYTES_IN_MB : undefined,
-      maxSizeJsonl: _.isFinite(opts.maxSizeJsonl)
+      maxSize: isFinite(opts.maxSize) ? Math.floor(opts.maxSize) * BYTES_IN_MB : undefined,
+      maxSizeJsonl: isFinite(opts.maxSizeJsonl)
         ? Math.floor(opts.maxSizeJsonl) * BYTES_IN_MB
         : undefined,
     },
@@ -93,71 +75,33 @@ module.exports = async (opts) => {
   const engine = createTransferEngine(source, destination, engineOptions);
 
   try {
-    let resultData = [];
     logger.log(`Starting export...`);
 
-    engine.progress.stream.on('start', ({ stage }) => {
+    // eslint-disable-next-line no-unused-vars
+    engine.progress.stream.on('start', ({ stage, data }) => {
       logger.log(`Starting transfer of ${stage}...`);
     });
 
-    // engine.progress.stream..on('progress', ({ stage, data }) => {
-    //   logger.log('progress');
+    // engine.progress.stream.on('progress', ({ stage, data }) => {
+    //   logger.log('progress', stage, data);
     // });
 
+    // eslint-disable-next-line no-unused-vars
     engine.progress.stream.on('complete', ({ stage, data }) => {
       logger.log(`...${stage} complete`);
-      resultData = data;
     });
 
     const results = await engine.transfer();
-
-    // Build pretty table
-    const table = new Table({
-      head: ['Type', 'Count', 'Size'],
-    });
-
-    let totalBytes = 0;
-    let totalItems = 0;
-    Object.keys(resultData).forEach((key) => {
-      const item = resultData[key];
-
-      table.push([
-        { hAlign: 'left', content: chalk.bold(key) },
-        { hAlign: 'right', content: item.count },
-        { hAlign: 'right', content: `${readableBytes(item.bytes, 1, 11)} ` },
-      ]);
-      totalBytes += item.bytes;
-      totalItems += item.count;
-
-      if (item.aggregates) {
-        Object.keys(item.aggregates).forEach((subkey) => {
-          const subitem = item.aggregates[subkey];
-
-          table.push([
-            { hAlign: 'left', content: `-- ${chalk.bold(subkey)}` },
-            { hAlign: 'right', content: subitem.count },
-            { hAlign: 'right', content: `(${chalk.grey(readableBytes(subitem.bytes, 1, 11))})` },
-          ]);
-        });
-      }
-    });
-    table.push([
-      { hAlign: 'left', content: chalk.bold.green('Total') },
-      { hAlign: 'right', content: chalk.bold.green(totalItems) },
-      { hAlign: 'right', content: `${chalk.bold.green(readableBytes(totalBytes, 1, 11))} ` },
-    ]);
+    const table = buildTransferTable(results.engine);
     logger.log(table.toString());
 
-    // TODO: once archiving is implemented, we need to check file extensions
     if (!fs.pathExistsSync(results.destination.file.path)) {
       logger.log(file);
       throw new Error('Export file not created');
     }
 
-    logger.log(`
-${chalk.bold('Export process has been completed successfully!')}
-Export archive is in ${chalk.green(results.destination.file.path)}
-`);
+    logger.log(`${chalk.bold('Export process has been completed successfully!')}`);
+    logger.log(`Export archive is in ${chalk.green(results.destination.file.path)}`);
     process.exit(0);
   } catch (e) {
     logger.error('Export process failed unexpectedly:', e.toString());
